Moved SyncSettings to core.config package.
